Voronoi Diagram from Data Frame

Taken from https://cran.r-project.org/src/contrib/Archive/ggvoronoi/ggvoronoi_0.8.4.tar.gz directly since it is a standalone function and the dependency on ggvoronoi is messing up R CMD CHECK

voronoi_polygon(data, x = "x", y = "y", outline = NULL, data.frame = FALSE)

Arguments

data

data.frame containing a set of points to make voronoi regions and any additional desired columns.

x

numeric vector (for example longitude).

y

numeric vector (for example latitude).

outline

data.frame with first column x/longitude, second column y/latitude, and a group column denoting islands or pieces.

data.frame

output as data.frame? You will lose information if you do this. For use in StatVoronoi.

Details

Create a Voronoi diagram for analysis or plotting with geom_polygon.

Examples

set.seed(45056)
x <- sample(1:200,100)
y <- sample(1:200,100)
points <- data.frame(x, y,
                     distance = sqrt((x-100)^2 + (y-100)^2))
circle <- data.frame(x = 100*(1+cos(seq(0, 2*pi, length.out = 2500))),
                     y = 100*(1+sin(seq(0, 2*pi, length.out = 2500))),
                     group = rep(1,2500))

vor_spdf <- voronoi_polygon(data=points,x="x",y="y",outline=circle)
#> Loading required package: sp
#> Loading required package: deldir
#> deldir 1.0-6      Nickname: "Mendacious Cosmonaut"
#> 
#>      The syntax of deldir() has had an important change. 
#>      The arguments have been re-ordered (the first three 
#>      are now "x, y, z") and some arguments have been 
#>      eliminated.  The handling of the z ("tags") 
#>      argument has been improved.
#>  
#>      The "dummy points" facility has been removed. 
#>      This facility was a historical artefact, was really 
#>      of no use to anyone, and had hung around much too 
#>      long.  Since there are no longer any "dummy points", 
#>      the structure of the value returned by deldir() has 
#>      changed slightly.  The arguments of plot.deldir() 
#>      have been adjusted accordingly; e.g. the character 
#>      string "wpoints" ("which points") has been 
#>      replaced by the logical scalar "showpoints". 
#>      The user should consult the help files.
vor_df <- fortify_voronoi(vor_spdf)
#> Error in fortify_voronoi(vor_spdf): could not find function "fortify_voronoi"

ggplot(vor_df) +
    geom_polygon(aes(x=x,y=y,fill=distance,group=group))
#> Error in ggplot(vor_df): could not find function "ggplot"